refactor: merge electron and server folders
