+++ server/package.json
@@ -6,7 +6,7 @@
     "name": "Narwhal Technologies Inc",
     "email": "hello@nrwl.io"
   },
-  "main": "electron.js",
+  "main": "src/index.js",
   "license": "MIT",
   "dependencies": {
     "@angular/cli": "6.1.2",
