+++ server/tsconfig.json
@@ -7,9 +7,9 @@
     "declaration": false,
     "moduleResolution": "node",
     "target": "es2017",
-    "typeRoots": ["../node_modules/@types"],
     "types": ["jasmine"],
-    "lib": ["es2017", "esnext.asynciterable"]
+    "typeRoots": ["../node_modules/@types"],
+    "lib": ["dom", "es2017", "esnext.asynciterable"]
   },
   "exclude": ["node_modules", "tmp"]
 }
