+++ tools/scripts/after-pack.js
@@ -4,10 +4,10 @@ const os = require('os');
 exports.default = async function(context) {
   if (context.appOutDir.indexOf('win-unpacked') > -1) {
     if (os.platform() === 'win32') {
-      execSync('copy tools\\win\\.bin\\ng.cmd dist\\electron\\ng.cmd');
+      execSync('copy tools\\win\\.bin\\ng.cmd dist\\server\\ng.cmd');
       execSync('copy tools\\win\\.bin\\ng.cmd dist\\packages\\win-unpacked\\resources\\app\\ng.cmd');
     } else {
-      execSync('cp tools/win/.bin/ng.cmd dist/electron/ng.cmd');
+      execSync('cp tools/win/.bin/ng.cmd dist/server/ng.cmd');
       execSync('cp tools/win/.bin/ng.cmd dist/packages/win-unpacked/resources/app/ng.cmd');
     }
   }
